# xe: sdpa: improve performance of quantized sdpa with a head size of 64