# Description

This PR updates the configurations for the SDPA kernel to optimized for a head size of 64. These updates improve the performance of small head sizes by 1.1-1.5x on LNL. Performance in other platforms will be posted soon.

```
| mb |  N |  D |   KV |    Q | kdt       | vdt       | mask   | quant                 | sdpa(main) | sdpa(PR) | speedup vs. main |
|----+----+----+------+------+-----------+-----------+--------+-----------------------+------------+----------+------------------|
|  1 | 32 | 64 |  385 |    1 | s8/f16/na | s8/f16/na | causal | per_token_with_groups |      80.75 |     62.5 |            1.292 |
|  1 | 32 | 64 |  513 |    1 | s8/f16/na | s8/f16/na | causal | per_token_with_groups |      91.96 |    70.78 |        1.2992371 |
|  1 | 32 | 64 | 1025 |    1 | s8/f16/na | s8/f16/na | causal | per_token_with_groups |     122.32 |   101.19 |        1.2088151 |
|  1 | 32 | 64 | 2049 |    1 | s8/f16/na | s8/f16/na | causal | per_token_with_groups |     216.77 |   154.16 |        1.4061365 |
|  1 | 32 | 64 | 4097 |    1 | s8/f16/na | s8/f16/na | causal | per_token_with_groups |     398.24 |    257.9 |        1.5441644 |
|----+----+----+------+------+-----------+-----------+--------+-----------------------+------------+----------+------------------|
|  1 | 32 | 64 |  384 |  384 | s8/f16/na | s8/f16/na | causal | per_token_with_groups |     402.67 |   299.89 |        1.3427257 |
|  1 | 32 | 64 |  512 |  512 | s8/f16/na | s8/f16/na | causal | per_token_with_groups |     512.16 |   458.93 |        1.1159872 |
|  1 | 32 | 64 | 1024 | 1024 | s8/f16/na | s8/f16/na | causal | per_token_with_groups |    1801.07 |  1616.98 |        1.1138480 |
|  1 | 32 | 64 | 2048 | 2048 | s8/f16/na | s8/f16/na | causal | per_token_with_groups |    9421.45 |  6120.72 |        1.5392715 |
|  1 | 32 | 64 | 4096 | 4096 | s8/f16/na | s8/f16/na | causal | per_token_with_groups |    26456.8 |  23432.6 |        1.1290595 |
#+TBLFM: $12=$10/$11
```

Addresses: [MFDNN-11755](https://jira.devtools.intel.com/browse/MFDNN-11755)

This PR also refactors the checks for the input descriptors so that its uniform between the internal primitive and the Graph API.